requirements.txt
dask
pandas
torchtext
pytorch-lightning
flake8
black
dgl
hydra-core
